Create new Shiftables module
